scripts/requirements.txt
black==24.8.0
isort==5.13.2
pre-commit==3.8.0
python-dotenv==1.0.1
PyYAML==6.0.2
